lab5 : complete

+++ labcodes_answer/lab5/libs/defs.h
@@ -10,7 +10,7 @@
 #define __noreturn __attribute__((noreturn))
 
 /* Represents true-or-false values */
-typedef int bool;
+typedef long long bool;
 
 /* Explicitly-sized versions of integer types */
 typedef char int8_t;
